LayoutTests/inspector/css/setLayoutContextTypeChangedMode-expected.txt

Tests for the CSS.setLayoutContextTypeChangedMode command.


== Running test suite: CSS.setLayoutContextTypeChangedMode
-- Running test case: CSS.setLayoutContextTypeChangedMode.queryGrid
PASS: 0 grid nodes should be instrumented.
Querying document for selector `#queryGrid`...
PASS: 1 grid node should be instrumented.
Changing `#queryGrid` to `display: block;`...
PASS: 0 grid nodes should be instrumented.
Changing `#queryGrid` to `display: grid;`...
PASS: 1 grid node should be instrumented.
Changing `#queryGrid` to `display: block;`...
PASS: 0 grid nodes should be instrumented.
Changing `#queryGrid` to `display: inline-grid;`...
PASS: 1 grid node should be instrumented.

-- Running test case: CSS.setLayoutContextTypeChangedMode.normalGrid
PASS: 1 grid node should be instrumented.
Changing `layoutContextTypeChangedMode` to `All`...
PASS: 2 grid nodes should be instrumented.
Changing `layoutContextTypeChangedMode` to `Observed`...
PASS: 2 grid nodes should be instrumented.
Changing `#normalGrid` to `display: block;`...
PASS: 1 grid node should be instrumented.
Changing `#normalGrid` to `display: grid;`...
PASS: 2 grid nodes should be instrumented.

-- Running test case: CSS.setLayoutContextTypeChangedMode.normalNonGrid
PASS: 2 grid nodes should be instrumented.
Changing `#normalNonGrid` to `display: grid;`...
PASS: 2 grid nodes should be instrumented.
Changing `#normalNonGrid` to `display: block;`...
PASS: 2 grid nodes should be instrumented.
Changing `layoutContextTypeChangedMode` to `All`...
PASS: 2 grid nodes should be instrumented.
Changing `#normalNonGrid` to `display: grid;`...
PASS: 3 grid nodes should be instrumented.
Changing `#normalNonGrid` to `display: block;`...
PASS: 2 grid nodes should be instrumented.
